Created new context for profile related components and created nested routes
